some control interface changes. added a small write up on the control interfaces

+++ src/main/java/frc/robot/Controls/ControlInterface.java
@@ -5,7 +5,7 @@ public interface ControlInterface {
     double getDriveY();
     double getDriveRot();
     boolean getSlowMode();
-    double getRobotRelativeDegrees();
+    int getRobotRelativeDegrees();
     boolean getShooterButton();
     boolean getIntakeButton();
     double getAngleAdjust();
